fix test

+++ featuretools/tests/primitive_tests/test_upper_case_word_count.py
@@ -19,9 +19,9 @@ class TestUpperCaseWordCount:
             dtype="string",
         )
         primitive_func = self.primitive().get_function()
-        answers = pd.Series([2, 1, 2, 3, 1], dtype='Int64')
+        answers = pd.Series([2, 1, 2, 3, 1], dtype="Int64")
         pd.testing.assert_series_equal(
-            primitive_func(x),
+            primitive_func(x).astype("Int64"),
             answers,
             check_names=False,
         )
@@ -36,9 +36,9 @@ class TestUpperCaseWordCount:
             dtype="string",
         )
         primitive_func = self.primitive().get_function()
-        answers = pd.Series([pd.NA, 0, 2], dtype='Int64')
+        answers = pd.Series([pd.NA, 0, 2], dtype="Int64")
         pd.testing.assert_series_equal(
-            primitive_func(x),
+            primitive_func(x).astype("Int64"),
             answers,
             check_names=False,
         )
